pyproject.toml
[project]
name = "image-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "pillow>=12.1.1",
]
